3.6.3

+++ src/services/starboard.ts
@@ -72,12 +72,11 @@ export const service: MessageService = {
                 const msg = m.message;
                 if (!msg.partial && msg.guild && msg.channel instanceof GuildChannel) {
                     const starboard = await client.database.getStarboardSetting(msg.guild.id);// getting the starboard settings
-                    const count = (m.count || 0);
-                    console.log(m.emoji)
+                    const count = (m.users.cache.filter(u => u.id !== client.user?.id).size);// try to get an accurate number of users who reacted with the current emoji
+                    // console.log("emoji:", m.emoji.id ?? m.emoji.name ?? "")
                     if (
                         !starboard.locked &&
                         starboard.channel &&
-                        count >= starboard.threshold &&
                         (starboard.allowSelf || msg.author.id !== user.id) &&
                         starboard.emoji.length &&
                         (starboard.emoji.find((e) => {
@@ -92,44 +91,54 @@ export const service: MessageService = {
                             return false;
                         })) &&
                         !starboard.ignoredChannels.includes(msg.channel.id)
-                    ) {// if the channel isn't locked, the id is present and not an empty string, the threshold is met, and the emoji is the right emoji
-                        console.log("past emoji check")
+                    ) {// if the channel isn't locked, the id is present and not an empty string, the threshold is met, and the emoji is the right emoji=
+                        // console.log("past emoji check")
                         const mid = msg.id;
                         const starPost = await client.database.getStarredMessage(mid);
                         const starChannel = msg.guild.channels.cache.get(starboard.channel);
                         const nsfw = !starPost ? msg.channel.nsfw : !!starPost.nsfw;
                         if (starboard.allowSensitive || !nsfw) {
-                            if (!starPost) {// if there is already a starboard entry for this message
-                                let postid = "";
-                                if (starChannel && starChannel.isText() && client.user && starChannel.permissionsFor(client.user)?.has([Permissions.FLAGS.SEND_MESSAGES, Permissions.FLAGS.EMBED_LINKS, Permissions.FLAGS.ATTACH_FILES])) {
-                                    const mdat = makeStarpost(starboard, msg, nsfw, count);
-                                    const post = await starChannel.send(mdat);// post a new entry to the starboard channel
-                                    if (starboard.starStarred) {
-                                        try {
-                                            await post.react(starboard.emoji[0]);
-                                        } catch (error) {
-                                            //
+                            if (!starPost || !starPost.postid) {// if there is already a starboard entry for this message
+                                if (count >= starboard.threshold) {// if the post has the necessary number of reactions
+                                    let postid = "";
+                                    if (starChannel && starChannel.isText() && client.user && starChannel.permissionsFor(client.user)?.has([Permissions.FLAGS.SEND_MESSAGES, Permissions.FLAGS.EMBED_LINKS, Permissions.FLAGS.ATTACH_FILES])) {
+                                        const mdat = makeStarpost(starboard, msg, nsfw, count);
+                                        const post = await starChannel.send(mdat);// post a new entry to the starboard channel
+                                        if (starboard.starStarred) {
+                                            try {
+                                                await post.react(starboard.emoji[0]);
+                                            } catch (error) {
+                                                //
+                                            }
                                         }
+                                        postid = post.id;
                                     }
-                                    postid = post.id;
+                                    await client.database.setStarredMessage({// save new entry no matter what, so it can be retrieved for whatever reason in the future
+                                        messageid: msg.id,
+                                        authorid: msg.author.id,
+                                        channelid: msg.channel.id,
+                                        guildid: msg.guild.id,
+                                        stars: count,
+                                        locked: 0,
+                                        nsfw: nsfw ? 1 : 0,
+                                        postid,
+                                        postchannel: starChannel?.id ?? "",
+                                    });
                                 }
-                                await client.database.setStarredMessage({// save new entry no matter what, so it can be retrieved for whatever reason in the future
-                                    messageid: msg.id,
-                                    authorid: msg.author.id,
-                                    channelid: msg.channel.id,
-                                    guildid: msg.guild.id,
-                                    stars: count,
-                                    locked: 0,
-                                    nsfw: nsfw ? 1 : 0,
-                                    postid,
-                                    postchannel: starChannel?.id ?? "",
-                                });
                             } else {// if an entry may need to be made
-                                if (starChannel && starChannel.isText() && client.user && starChannel.permissionsFor(client.user)?.has([Permissions.FLAGS.SEND_MESSAGES, Permissions.FLAGS.EMBED_LINKS, Permissions.FLAGS.ATTACH_FILES])) {
-                                    if (starChannel.id === starPost.postchannel && isSnowflake(starPost.postid) && count !== starPost.stars) {
+                                // console.log("found existing message")
+                                if (starChannel && starChannel.isText() && client.user && starChannel.permissionsFor(client.user)?.has([Permissions.FLAGS.SEND_MESSAGES, Permissions.FLAGS.EMBED_LINKS, Permissions.FLAGS.ATTACH_FILES])) {// the starboard channel is set and messages can be sent in it
+                                    // console.log("passed msg perm check:", starPost.postid)
+                                    if (starChannel.id === starPost.postchannel && isSnowflake(starPost.postid)/*  && count !== starPost.stars */) {// checks to make sure the postid is valid and that the channel of the post is in the starboard channel, since it may have changed
+                                        // console.log("secondary check:", count)
                                         if (count === 0) {
-                                            const post = await starChannel.messages.fetch(starPost.postid);
-                                            if (post.deletable) {
+                                            const post = await starChannel.messages.fetch(starPost.postid).catch(() => {
+                                                starPost.postchannel = "";
+                                                starPost.postid = "";
+                                            });
+                                            if (post && post.deletable) {
+                                                starPost.postid = "";
+                                                starPost.postchannel = "";
                                                 await post.delete();
                                             }
                                         } else {
@@ -137,6 +146,10 @@ export const service: MessageService = {
                                             const post = await starChannel.messages.fetch(starPost.postid);
                                             await post.edit(mdat);// edit the starpost for posterity (to see an updated count)
                                         }
+                                    } else {// the location of the starpost is in the wrong channel, or the post id was invalid
+                                        // the id and channel should be reset so it can be reposted
+                                        starPost.postid = "";
+                                        starPost.postchannel = "";
                                     }
                                 }
                                 await client.database.setStarredMessage({// update stored star count
